refactor: restructured utility functions
Broke up common utils shared across packages into separate modules within a larger "utitilities" package
