+++ src/constants/sources.py
@@ -6,7 +6,7 @@ Constants for various sources (e.g. file paths, URLs).
 TODO: Consider renaming to `paths.py`.
 """
 
-from src.utils import find_project_root
+from src.utilities.common import find_project_root
 
 ### --- CONSTANTS --- ###
 # OHA Data Book
