+++ src/utilities/__init__.py
@@ -1,0 +1,5 @@
+"""
+Utilities
+
+Utility functions that are used across the project and are not specific to any one module (e.g. llmcore, datacore, etc.).
+"""
